refactored common code for creating temporary folder from KafkaUnit to LogDirUtil, make use of it also from ZooKeeper
